requirements.txt
absl-py==0.7.1
annoy==1.15.1
arrow==0.13.1
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.6
bert-serving-client==1.8.6
bert-serving-server==1.8.6
bleach==3.1.0
blis==0.2.4
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
decorator==4.4.0
defusedxml==0.5.0
Django==2.0.4
djangorestframework==3.10.3
elephas==0.4.2
entrypoints==0.3
Flask==1.1.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.19.0
h5py==2.9.0
hyperas==0.4.1
hyperopt==0.1.2
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-bert==0.80.0
keras-embed-sim==0.7.0
keras-layer-normalization==0.13.0
keras-multi-head==0.22.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.0
keras-self-attention==0.41.0
keras-transformer==0.31.0
kiwisolver==1.1.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mmh3==2.5.1
murmurhash==1.0.2
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
notebook==5.7.6
numpy==1.16.2
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
py4j==0.10.7
pyasn1==0.4.5
pycparser==2.19
pydot==1.4.1
Pygments==2.3.1
pymongo==3.7.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.0
pyrsistent==0.14.11
pyspark==2.4.4
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
spacy==2.1.0
srsly==0.0.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==7.0.4
tornado==6.0.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wasabi==0.1.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.0
Whoosh==2.7.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xgboost==0.90
